+++ dwds/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 24.3.3-wip
-- Added support for some debugging APIs with the DDC library bundle format. - [#2563](https://github.com/dart-lang/webdev/issues/2563)
+- Added support for some debugging APIs with the DDC library bundle format. - [#2563](https://github.com/dart-lang/webdev/issues/2563),[#2566](https://github.com/dart-lang/webdev/issues/2566)
 
 ## 24.3.2